requirements.txt
# --- Core & Framework ---
Django==6.0.2
djangorestframework==3.16.1    
django-cors-headers==4.9.0     
django-filter==25.2
python-dotenv==1.2.2
django-extensions==3.2.3
django-silk==5.3.2
drf-spectacular==0.28.0
djoser==2.3.1

# --- Database ---
psycopg2-binary==2.9.10
sqlparse==0.5.5

# --- Utils & Dependencies ---
PyJWT==2.11.0
asgiref==3.8.1
requests==2.32.5
urllib3==2.6.3
tzdata==2025.3
python-dateutil==2.9.0.post0
inflection==0.5.1
jsonschema==4.26.0
pyparsing==3.3.2
certifi==2025.1.31
idna==3.10
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
defusedxml==0.7.1
stripe==11.5.0
celery==5.4.0
redis==5.2.1
django-storages==1.14.4
boto3==1.34.162
django-cleanup==9.0.0
eventlet==0.36.1
